movinr exceptions
