+++ cnxarchive/views.py
@@ -410,53 +410,6 @@ def _get_licenses(cursor):
 
 
 
-# ################### #
-#   Exception Views   #
-# ################### #
-
-
-@view_config(context=IdentHashSyntaxError)
-def ident_hash_syntax_error(exc, request):
-    if exc.ident_hash.endswith('@'):
-        try:
-            split_ident_hash(exc.ident_hash[:-1])
-        except IdentHashShortId as e:
-            return ident_hash_short_id(e, request)
-        except IdentHashMissingVersion as e:
-            return ident_hash_missing_version(e, request)
-        except IdentHashSyntaxError:
-            pass
-    return httpexceptions.HTTPNotFound()
-
-
-@view_config(context=IdentHashShortId)
-def ident_hash_short_id(exc, request):
-    try:
-        uuid_ = get_uuid(exc.id)
-    except httpexceptions.HTTPNotFound as e:
-        return e
-    if not exc.version:
-        return ident_hash_missing_version(
-            IdentHashMissingVersion(uuid_), request)
-    route_name = request.matched_route.name
-    route_args = request.matchdict.copy()
-    route_args['ident_hash'] = join_ident_hash(uuid_, exc.version)
-    return httpexceptions.HTTPFound(request.route_path(
-        route_name, _query=request.params, **route_args))
-
-
-@view_config(context=IdentHashMissingVersion)
-def ident_hash_missing_version(exc, request):
-    try:
-        version = get_latest_version(exc.id)
-    except httpexceptions.HTTPNotFound as e:
-        return e
-    route_name = request.matched_route.name
-    route_args = request.matchdict.copy()
-    route_args['ident_hash'] = join_ident_hash(exc.id, version)
-    return httpexceptions.HTTPFound(request.route_path(
-        route_name, _query=request.params, **route_args))
-
 
 # ######### #
 #   Views   #
